Setup github authentication on the jenkins server
Setup and configure github based authentication on the jenkins server. Render the scripts and authentication files and configure access rights and user roles in the jenkins server.

For the first run, when the authentication is not enabled yet, don't use any authentication mechanisms (implied by api_token being false). Afterwards, configure the token and it will be used for the authentication.