Wrapper script for vault password
On Mac OS X this will attempt to fetch the password from the keychain.
This is more secure than storing the password in a plaintext file on
the filesystem.

On Linux it will fallback to an interactive prompt which is effectively
the same as `--ask-vault-pass` but allows us to simplify our run
instructions by using the same argument regardless.

When the argument to `--vault-password-file` is an executable then
Ansible will run it an obtain the password from STDOUT.