Define host external IP based on real value
Define external IP of the jenkins server, which is later used for waiting
on server and configuring DNS record, to the correct value - based on
whether Elastic IP is associated or not. Before, this was defined as
server's external IP (ec2.tagged_instances[0].public_ip) - which was
correct, except the variable was registered before Elastic IP association,
hence containing old IP. To configure the correct IP in the DNS record,
you would have to run the playbook twice - the second time you'd read
the correct external IP, as this was changed to EIP in the previous run.
This fixes the Elastic IP handlig part. If you didn't use Elastic IPs,
the playbook worked correctly the 1st time.