Add job definitions for smoke tests on AWS and GCE
We add the definitions to add the jobs to run the smoke tests
against each one of the platforms: AWS and GCE.

Due limitations in ansible, which does not allow override variables for templates,
we were force to create two templates to keep the ansible logic simpler. A better
solution can be add some custom resources or a role to define jenkins jobs.

The jobs depend on the jenkins job that applies ansible on each environment.

Note: There is no job locking or similar that prevents this job to run while the
environment is deployed or provisioned. We would need to ad this logic if required.