Create enviroment specific terraform statefiles
__Problem__ :
We are using a single statefile to handle multiple environments and that doesn't work because
you can only target one environment at a time.

__Observed outputs__ :
Previously when deploying environments, the deployments would overwrite
the single terrafrom statefile, causing the deployment to not reach a desired
outcome even if the job succeeds.

__Solution__ :
This commit adds a flag to the command line running of Terraform to specify a environment
terraform state file.