suspend env. jobs: Don't fail when nothing to do
When all the environments to be suspended are excluded, grep exits with 1, which is caught by the `set -e`. The job fails, but should succeed as there is nothing to do and all the environments that are supposed to be suspended (in this specific case none) are suspended. Fix this by adding `|| echo` after grep. This way job will not fail and the rest of the script will try to suspend "" environments. Which it will succeed.