Vagrant: disable email notifications
Currently all email notifications are enabled on all jobs. In vagrant,
we disable the jobs. But we don't disable or remove the notifications.
When we then develop jobs (which means they get enabled), we get spam
in the main team mailbox for every local vagrant issue. Disable email
notifications for Vagrant.