Fix the dashboard so that it fills the whole screen.
This was set to 1000px high, so was leaving a blank space at the bottom
on a 1080p display. Setting this to 1064px allows it to fill the whole
screen (allowing for the 8px padding top and bottom).

The 8px margin was being added by the Safari user-agent stylesheet, so I
added it here to make it explicit, and also make it more obvious why the
height is set to 1064px.